% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case1_pkpd}
\alias{case1_pkpd}
\title{Case 1 PKPD Data Set}
\format{A data frame with the following 21 columns:
\tabular{rll}{
  column 1: \tab \code{ID} \tab integer; unique subject ID\cr
  column 2: \tab \code{TIME} \tab numeric; time relative to first drug 
  administration\cr
  column 3: \tab \code{NOMTIME} \tab numeric; nominal time\cr
  column 4: \tab \code{TIMEUNIT} \tab factor; unit of TIME\cr
  column 5: \tab \code{AMT} \tab integer; dosing amount (for dosing events) 
  in mg\cr
  column 6: \tab \code{LIDV} \tab numeric; observation on a linear scale 
  (observation type determined by CMT), units determined by EVENTU column\cr
  column 7: \tab \code{CMT} \tab integer; compartment number 
  (determines observation type):\cr
  \tab \tab CMT 1 = Dosing event\cr
  \tab \tab CMT 2 = PK concentration\cr
  \tab \tab CMT 3 = Continuous response data\cr
  \tab \tab CMT 4 = Count response data\cr
  \tab \tab CMT 5 = Ordinal response data\cr
  \tab \tab CMT 6 = Binary response data\cr
  column 8: \tab \code{NAME} \tab factor; description of event\cr
  column 9: \tab \code{EVENTU} \tab factor; unit for observation\cr
  column 10: \tab \code{CENS} \tab integer; censored values
  (0 = not censored, 1 = censored)\cr
  column 11: \tab \code{EVID} \tab integer; event ID (0 = observation, 
  1 = dosing event)\cr
  column 12: \tab \code{WEIGHTB} \tab numeric; baseline body weight (kg)\cr
  column 13: \tab \code{eff0} \tab numeric; efficacy\cr
  column 14: \tab \code{TRTACT} \tab factor; treatment group label\cr
  column 15: \tab \code{DOSE} \tab integer; Dose in mg\cr
  column 16: \tab \code{PROFDAY} \tab integer; day of profile\cr
  column 17: \tab \code{PROFTIME} \tab numeric; time within PROFDAY\cr
  column 18: \tab \code{CYCLE} \tab integer; count of drug administrations
  received\cr
  column 19: \tab \code{PART} \tab integer; part of study\cr
  column 20: \tab \code{STUDY} \tab integer; study\cr
  column 21: \tab \code{IPRED} \tab numeric; individual prediction\cr
}}
\usage{
case1_pkpd
}
\description{
Case 1 PKPD Data Set
}
\keyword{datasets}
