% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mad_missing_duplicates}
\alias{mad_missing_duplicates}
\title{Multiple Ascending Dose Data Set (Duplicates Removed)}
\format{
A data frame with the following 19 columns:
\tabular{rll}{
  column 1: \tab \code{ID} \tab numeric; unique subject ID\cr
  column 2: \tab \code{TIME} \tab numeric; time relative to first drug 
  administration\cr
  column 3: \tab \code{NOMTIME} \tab numeric; nominal time\cr
  column 4: \tab \code{TIMEUNIT} \tab character; unit of TIME\cr
  column 5: \tab \code{AMT} \tab numeric; dosing amount (for dosing events) 
  in mg\cr
  column 6: \tab \code{LIDV} \tab numeric; observation on a linear scale 
  (observation type determined by CMT), units determined by EVENTU column\cr
  column 7: \tab \code{MDV} \tab numeric; missing dependent variable\cr
  column 8: \tab \code{CMT} \tab integer; compartment number 
  (determines observation type):\cr
  \tab \tab CMT 1 = Dosing event\cr
  \tab \tab CMT 2 = PK concentration\cr
  \tab \tab CMT 3 = Continuous response data\cr
  \tab \tab CMT 4 = Count response data\cr
  \tab \tab CMT 5 = Ordinal response data\cr
  \tab \tab CMT 6 = Binary response data\cr
  column 9: \tab \code{NAME} \tab character; description of event\cr
  column 10: \tab \code{EVENTU} \tab character; unit for observation\cr
  column 11: \tab \code{CENS} \tab integer; censored values
  (0 = not censored, 1 = censored)\cr
  column 12: \tab \code{EVID} \tab integer; event ID (0 = observation, 
  1 = dosing event)\cr
  column 13: \tab \code{WEIGHTB} \tab numeric; baseline body weight (kg)\cr
  column 14: \tab \code{SEX} \tab character; sex\cr
  column 15: \tab \code{TRTACT} \tab factor; treatment group label\cr
  column 16: \tab \code{DOSE} \tab numeric; randomized dose in mg\cr
  column 17: \tab \code{PROFDAY} \tab numeric; day of profile\cr
  column 18: \tab \code{PROFTIME} \tab numeric; time within PROFDAY\cr
  column 19: \tab \code{CYCLE} \tab numeric; count of drug administrations 
  received
}
}
\usage{
mad_missing_duplicates
}
\description{
Model generated PK and PD data to mimic an orally administered small
molecule with various endpoints from continuous to ordinal response and
count data. Simulated multiple dose administration ranging from 100 mg
to 1600 mg, once per day.
}
\keyword{datasets}
