% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlmixr_nimo}
\alias{nlmixr_nimo}
\title{nlmixr Data Set}
\format{A data frame with the following 20 columns:
\tabular{rll}{
  column 1: \tab \code{ID} \tab integer; TODO\cr
  column 2: \tab \code{TIME} \tab numeric; TODO\cr
  column 3: \tab \code{AMT} \tab integer; TODO\cr
  column 4: \tab \code{RATE} \tab numeric; TODO\cr
  column 5: \tab \code{DV} \tab numeric; TODO\cr
  column 6: \tab \code{TAD} \tab numeric; TODO\cr
  column 7: \tab \code{OCC} \tab integer; TODO\cr
  column 8: \tab \code{FLAG} \tab integer; TODO\cr
  column 9: \tab \code{MDV} \tab integer; TODO\cr
  column 10: \tab \code{EVID} \tab integer; TODO\cr
  column 11: \tab \code{WGT} \tab integer; TODO\cr
  column 12: \tab \code{BSA} \tab numeric; TODO\cr
  column 13: \tab \code{AGE} \tab integer; age in years\cr
  column 14: \tab \code{RACE} \tab integer; TODO\cr
  column 15: \tab \code{HGT} \tab integer; TODO\cr
  column 16: \tab \code{DOS} \tab integer; TODO\cr
  column 17: \tab \code{STR} \tab integer; TODO\cr
  column 18: \tab \code{INIT} \tab integer; TODO\cr
  column 19: \tab \code{CONC} \tab numeric; TODO\cr
  column 20: \tab \code{hipoalbuminemia} \tab integer; TODO
}}
\usage{
nlmixr_nimo
}
\description{
TODO
}
\keyword{datasets}
