% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlmixr_theo_sd}
\alias{nlmixr_theo_sd}
\title{NLMIXR Theo SD Data Set}
\format{A data frame with the following 7 columns:
\tabular{rll}{
  column 1: \tab \code{ID} \tab integer; TODO\cr
  column 2: \tab \code{TIME} \tab numeric; TODO\cr
  column 3: \tab \code{DV} \tab numeric; TODO\cr
  column 4: \tab \code{AMT} \tab numeric; TODO\cr
  column 5: \tab \code{EVID} \tab integer; TODO\cr
  column 6: \tab \code{CMT} \tab integer; TODO\cr
  column 7: \tab \code{WT} \tab numeric; TODO
}}
\usage{
nlmixr_theo_sd
}
\description{
TODO
}
\keyword{datasets}
