% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_stat_smooth.R
\name{predictdf}
\alias{predictdf}
\title{Prediction data frame from ggplot2
Get predictions with standard errors into data frame}
\usage{
predictdf(model, xseq, se, level)
}
\arguments{
\item{model}{model object}

\item{xseq}{newdata}

\item{se}{Display confidence interval around smooth?}

\item{level}{Level of confidence interval to use}
}
\description{
Prediction data frame from ggplot2
Get predictions with standard errors into data frame
}
