% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sad_pkpd}
\alias{sad_pkpd}
\title{Single Ascending Dose PKPD Data Set}
\format{A data frame with the following 16 columns:
\tabular{rll}{
  column 1: \tab \code{ID} \tab numeric; unique subject ID\cr
  column 2: \tab \code{TIME} \tab numeric; TODO\cr
  column 3: \tab \code{COBS} \tab numeric; TODO\cr
  column 4: \tab \code{EOBS} \tab numeric; TODO\cr
  column 5: \tab \code{WEIGHT} \tab numeric; TODO\cr
  column 6: \tab \code{AGE} \tab numeric; TODO\cr
  column 7: \tab \code{DOSE} \tab integer; TODO\cr
  column 14: \tab \code{SEX} \tab character; sex\cr
  column 15: \tab \code{RACE} \tab character; TODO\cr
  column 16: \tab \code{AMT} \tab integer; TODO
}}
\usage{
sad_pkpd
}
\description{
TODO
}
\keyword{datasets}
