% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_stat_ci.R
\docType{data}
\name{StatCI}
\alias{StatCI}
\title{Stat ggproto object for creating ggplot layers of binned confidence intervals}
\format{An object of class \code{StatCI} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatCI
}
\value{
ggplot2 ggproto object
}
\description{
\code{StatCI} returns a ggproto object for plotting mean +/- confidence bins
}
\keyword{datasets}
