% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_stat_smooth.R
\name{StatSmoothOrdinal}
\alias{StatSmoothOrdinal}
\title{Stat object for producing smooths through ordinal data}
\usage{
StatSmoothOrdinal
}
\description{
Stat object for producing smooths through ordinal data
}
