% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_stat_ci.R
\name{StatSummaryBinQuant}
\alias{StatSummaryBinQuant}
\title{Stat ggproto object for binning by quantile for xgx_stat_ci}
\usage{
StatSummaryBinQuant
}
\value{
ggplot2 ggproto object
}
\description{
Source:
    https://github.com/tidyverse/ggplot2/blob/351eb41623397dea20ed0059df62a4a5974d88cb/R/stat-summary-bin.R
}
\details{
\code{StatSummaryBinQuant} returns a ggproto object for plotting mean +/- confidence bins
}
