% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_stat_ci.R
\name{StatSummaryOrdinal}
\alias{StatSummaryOrdinal}
\title{Stat ggproto object for creating ggplot layers of binned confidence intervals
for probabiliities of classes in ordinal data}
\usage{
StatSummaryOrdinal
}
\value{
ggplot2 ggproto object
}
\description{
\code{StatSummaryOrdinal} returns a ggproto object for plotting mean +/- confidence intervals
for ordinal data. It also allows for binning values on the independent axis.
}
