% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_trans_units.R
\name{xgx_1_unit}
\alias{xgx_1_unit}
\title{A dummy translation function to capture the units of the x-axis}
\usage{
xgx_1_unit(set_unit)
}
\arguments{
\item{set_unit}{When specified, this will save the unit information
for x-axis transformation.}
}
\description{
A dummy translation function to capture the units of the x-axis
}
