% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_trans_units.R
\name{xgx_log_trans_unit}
\alias{xgx_log_trans_unit}
\alias{xgx_log10_trans_unit}
\title{xgx log trans for units}
\usage{
xgx_log_trans_unit(base = exp(1))

xgx_log10_trans_unit()
}
\arguments{
\item{base}{base of logarithm}
}
\description{
This takes units and transforms them for a log10 axes
}
