% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_scale_y_percentchangelog10.R
\name{xgx_scale_y_percentchangelog10}
\alias{xgx_scale_y_percentchangelog10}
\title{percentchangelog10 transform for the y scale.}
\usage{
xgx_scale_y_percentchangelog10(breaks = NULL, minor_breaks = NULL,
  labels = NULL, accuracy = 1, ...)
}
\arguments{
\item{breaks}{if NULL, then default is to use xgx_breaks_log10(x + 1) - 1}

\item{minor_breaks}{if NULL, then defauls is to use nicely spaced log(PCHG + 1) minor breaks}

\item{labels}{if NULL, then the default is to use scales::percent()}

\item{accuracy}{if NULL, then the default is set to 1}

\item{...}{other parameters passed to 
\code{\link[ggplot2:scale_continuous]{scale_y_continuous}}}
}
\value{
ggplot2 compatible scale object
}
\description{
\code{xgx_scale_y_percentchangelog10} is designed to be used with 
percent change from baseline data (on a scale of -1 to +Inf)
Common examples include % weight loss, % reduction in LDL, % change in tumor diameter.  
It is used when you have a wide range of data on a percent change scale, 
especially data close to -100%, and/or several fold increase from baseline.
}
\examples{
 
  
Nsubj <- 10
Doses <- c(0, 25, 50, 100, 200)
Ntot <- Nsubj*length(Doses)
dat1 <- data.frame(ID = 1:(Ntot),
                  DOSE = rep(Doses, Nsubj),
                  PD0 = rlnorm(Ntot, log(100), 1),
                  Kout = exp(rnorm(Ntot,-2, 0.3)),
                  Imax = 1,
                  ED50 = 25) \%>\%
 dplyr::mutate(PDSS = PD0*(1 - Imax*DOSE/(DOSE + ED50))*exp(rnorm(Ntot, 0.05, 0.3))  ) \%>\%
 dplyr::mutate(PCHG = (PDSS - PD0)/PD0)

ggplot2::ggplot(dat1, ggplot2::aes(x = DOSE, y = PCHG, group = DOSE)) +
  ggplot2::geom_boxplot() +
  xgx_theme() +
  xgx_scale_y_percentchangelog10()

dat2 <- data.frame(ID = 1:(Ntot),
                  DOSE = rep(Doses, Nsubj),
                  PD0 = rlnorm(Ntot, log(100), 1),
                  Kout = exp(rnorm(Ntot,-2, 0.3)),
                  Emax = 50*rlnorm(Ntot, 0, 0.3),
                  ED50 = 300) \%>\%
 dplyr::mutate(PDSS = PD0*(1 + Emax*DOSE/(DOSE + ED50))*exp(rnorm(Ntot, -1, 0.3))  ) \%>\%
 dplyr::mutate(PCHG = (PDSS - PD0)/PD0)

ggplot2::ggplot(dat2, ggplot2::aes(x = DOSE, y = PCHG, group = DOSE)) +
  ggplot2::geom_boxplot() +
  xgx_theme() +
  xgx_scale_y_percentchangelog10()

  
 
}
