% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_stat_pi.R
\name{xgx_stat_pi}
\alias{xgx_stat_pi}
\title{Plot data with median and percent intervals}
\usage{
xgx_stat_pi(
  mapping = NULL,
  data = NULL,
  percent_level = 0.95,
  geom = list("line", "ribbon"),
  position = "identity",
  fun.args = list(),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by `aes` or `aes_`. 
If specified and `inherit.aes = TRUE` (the default), it is combined with the 
default mapping at the top level of the plot. You must supply mapping if 
there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified 
in the call to ggplot.

A data.frame, or other object, will override the plot data. All objects 
will be fortified to produce a data frame. See fortify for which variables 
will be created.

A function will be called with a single argument, the plot data. The return 
value must be a data.frame., and will be used as the layer data.}

\item{percent_level}{The upper or lower percentile for the percent interval (should fall 
between 0 and 1). The default is 0.95, which corresponds  
to (0.05, 0.95) interval. Supplying 0.05 would give the same result}

\item{geom}{Use to override the default geom. Can be a list of multiple 
geoms, e.g. list("line","ribbon"), which is the default.}

\item{position}{Position adjustment, either as a string, or the result of 
a call to a position adjustment function.}

\item{fun.args}{Optional additional arguments passed on to the functions.}

\item{na.rm}{If FALSE, the default, missing values are removed with a 
warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? 
NA, the default, includes if any aesthetics are mapped. FALSE never 
includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather 
than combining with them. This is most useful for helper functions that 
define both data and aesthetics and shouldn't inherit behaviour from the 
default plot specification, e.g. borders.}

\item{...}{other arguments passed on to layer. These are often aesthetics, 
used to set an aesthetic to a fixed value, like color = "red" or size = 3. 
They may also be parameters to the paired geom/stat.}
}
\value{
ggplot2 plot layer
}
\description{
\code{xgx_stat_pi} returns a ggplot layer plotting median +/- percent 
intervals
}
\examples{
# default settings for normally distributed data, (5\%,95\%) interval,
data <- data.frame(x = rep(c(1, 2, 3), each = 20),
                   y = rep(c(1, 2, 3), each = 20) + stats::rnorm(60),
                   group = rep(1:3, 20))
xgx_plot(data, ggplot2::aes(x = x, y = y)) +
  xgx_stat_pi(percent_level = 0.95)

# try different geom 
xgx_plot(data, ggplot2::aes(x = x, y = y)) + 
  xgx_stat_pi(percent_level = 0.95, geom = list("errorbar", "point", "line"))
 
# including multiple groups in same plot
xgx_plot(data, ggplot2::aes(x = x, y = y)) + 
  xgx_stat_pi(percent_level = 0.95, 
              ggplot2::aes(color = factor(group), fill = factor(group)),
              position = ggplot2::position_dodge(width = 0.5))
              
# including multiple percent intervals in same plot              
xgx_plot(data, ggplot2::aes(x = x, y = y)) +
  xgx_stat_pi(percent_level = 0.90) + 
  xgx_stat_pi(percent_level = 0.80) + 
  xgx_stat_pi(percent_level = 0.70) + 
  xgx_stat_pi(percent_level = 0.60)      
 
}
