% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_theme_set.R
\name{xgx_theme_set}
\alias{xgx_theme_set}
\title{Sets the standard theme for xGx graphics}
\usage{
xgx_theme_set()
}
\value{
xgx ggplot2 compatible theme
}
\description{
xgx_theme_set
}
\examples{
conc <- 10^(seq(-3, 3, by = 0.1))
EC50 <- 1
data <- data.frame(concentration = conc, 
                   bound_receptor = 1 * conc / (conc + EC50))
xgx_theme_set()
ggplot2::ggplot(data, ggplot2::aes(y = concentration, x = bound_receptor)) +
  ggplot2::geom_point() +
  ggplot2::geom_line() +
  xgx_scale_y_log10() +
  xgx_scale_x_reverselog10()
  
}
