% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_stat_smooth.R
\name{predictdf.polr}
\alias{predictdf.polr}
\title{Prediction data frame for polr}
\usage{
\method{predictdf}{polr}(model, xseq, se, level)
}
\arguments{
\item{model}{object returned from polr}

\item{xseq}{sequence of x values for which to compute the smooth}

\item{se}{if TRUE then confidence intervals are returned}

\item{level}{confidence level for confidence intervals}
}
\description{
Get predictions with standard errors into data frame for use with geom_smooth
}
\details{
\code{predictdf.polr} is used by xgx_geom_smooth when method = "polr"
to calculate confidence intervals via bootstraps.
}
