% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_stat_smooth.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{nlsLM}
\alias{gtable}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gtable}{\code{\link[gtable]{gtable}}}

  \item{minpack.lm}{\code{\link[minpack.lm]{nlsLM}}}
}}

