% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgx_stat_smooth.R
\docType{data}
\name{StatSmoothOrdinal}
\alias{StatSmoothOrdinal}
\title{Stat object for producing smooths through ordinal data}
\format{
An object of class \code{StatSmoothOrdinal} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 8.
}
\usage{
StatSmoothOrdinal
}
\description{
Stat object for producing smooths through ordinal data
}
\keyword{datasets}
